PulseAudio: fix scaling of latency values using PA_USEC_PER_SEC

+++ src/hostapi/pulseaudio/pa_linux_pulseaudio_internal.h
@@ -128,7 +128,7 @@ typedef struct PaPulseAudio_Stream
     pa_usec_t outStreamTime;
     pa_buffer_attr bufferAttr;
     int outputUnderflows;
-    int latency;
+    double latency;
     int outputChannelCount;
 
     int callbackMode;       /* bool: are we running in callback mode? */
